Overhauled authentication and split into multiple activities; finished popularity guessing game; fixed minor bug in genre fetching; various UI fixes
