+++ app/src/main/AndroidManifest.xml
@@ -13,12 +13,29 @@
         android:theme="@style/Theme.Project2"
         tools:targetApi="31">
         <activity
-            android:name=".AuthenticationActivity"
-            android:exported="true"
-            android:label="@string/app_name">
+            android:name=".ui.authentication.IntroActivity"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
+                <data android:host="auth" android:scheme="project2"/>
+            </intent-filter>
+        </activity>
+        <activity
+            android:name=".ui.authentication.CreateAccountActivity"
+            android:exported="true"
+            android:label="Create Account">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <data android:host="auth" android:scheme="project2"/>
+            </intent-filter>
+        </activity>
+        <activity
+            android:name=".ui.authentication.LoginActivity"
+            android:exported="true"
+            android:label="Log In">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
                 <data android:host="auth" android:scheme="project2"/>
             </intent-filter>
         </activity>
@@ -32,6 +49,7 @@
                 <data android:host="auth" android:scheme="project2"/>
             </intent-filter>
         </activity>
+
         <meta-data
             android:name="preloaded_fonts"
             android:resource="@array/preloaded_fonts" />
