+++ app/src/main/res/layout/fragment_popularitygame.xml
@@ -44,7 +44,7 @@
             android:id="@+id/popularity_game_text"
             android:layout_width="280dp"
             android:layout_height="wrap_content"
-            android:layout_marginTop="52dp"
+            android:layout_marginTop="36dp"
             android:fontFamily="@font/roboto_mono"
             android:text="popularity_game_text"
             android:textAlignment="center"
@@ -84,7 +84,8 @@
             android:text="popular_artist_one_text"
             android:textAlignment="center"
             android:textColor="#FFFFFF"
-            android:textSize="16sp"
+            android:textSize="18sp"
+            android:textStyle="bold"
             app:layout_constraintBottom_toTopOf="@+id/return_to_dashboard_from_popularity_button"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/popular_artist_image_one"
@@ -100,7 +101,8 @@
             android:text="TextView"
             android:textAlignment="center"
             android:textColor="#FFFFFF"
-            android:textSize="16sp"
+            android:textSize="18sp"
+            android:textStyle="bold"
             app:layout_constraintBottom_toTopOf="@+id/return_to_dashboard_from_popularity_button"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/popular_artist_image_two"
