+++ app/src/main/res/layout/fragment_song_release_date.xml
@@ -67,14 +67,14 @@
             android:id="@+id/return_to_dashboard_from_song_game_button"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginTop="40dp"
+            android:layout_marginTop="24dp"
             android:text="@string/return_to_dashboard"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.503"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/song_submit_guess"
-            app:layout_constraintVertical_bias="0.078" />
+            app:layout_constraintVertical_bias="0.0" />
 
         <ImageView
             android:id="@+id/song_image"
