+++ app/src/main/res/values/preloaded_fonts.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <array name="preloaded_fonts" translatable="false">
+        <item>@font/monofett</item>
         <item>@font/roboto_mono</item>
     </array>
 </resources>
